refact: Seperate controller endpoints - view를 반환하는 엔드포인트를 ProductViewController로 분리.
