+++ src/test/java/gift/controller/ProductControllerTest.java
@@ -1,4 +1,4 @@
-package gift;
+package gift.controller;
 
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.delete;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
@@ -26,20 +26,6 @@ class ProductControllerTest {
     private @Autowired MockMvc mockMvc;
 
     @Test
-    @DisplayName("상품 목록 가져오기 테스트")
-    void getProducts() throws Exception {
-        mockMvc.perform(get("/api/products"))
-            .andExpect(status().isOk());
-    }
-
-    @Test
-    @DisplayName("상품 추가 폼 페이지 가져오기 테스트")
-    void addProductForm() throws Exception {
-        mockMvc.perform(get("/api/products/product"))
-            .andExpect(status().isOk());
-    }
-
-    @Test
     @DisplayName("상품 추가 테스트")
     void addProduct() throws Exception {
         String requestJson = """
@@ -49,21 +35,6 @@ class ProductControllerTest {
         mockMvc.perform(post("/api/products/product")
                 .contentType(MediaType.APPLICATION_JSON)
                 .content(requestJson))
-            .andExpect(status().isOk());
-    }
-
-    @Test
-    @DisplayName("상품 수정 폼 페이지 테스트")
-    void editProductForm() throws Exception {
-        String requestJson = """
-            {"id": 10,"name": "커피", "price": 5500,"imageUrl": "https://..."}
-            """;
-
-        mockMvc.perform(post("/api/products/product")
-            .contentType(MediaType.APPLICATION_JSON)
-            .content(requestJson));
-
-        mockMvc.perform(get("/api/products/product/10"))
             .andExpect(status().isOk());
     }
